Switch to using ssh metadata instead of oslogin
I encountered lots of issues with oslogin, the biggest is that it forces
us to log into a new user where we really want to login as the `sdci`
user (which has all the vagrant+libvirt settings applied already).

In comparison, passing metadata on instance creation is super
straight-forward. Followed docs from
https://cloud.google.com/compute/docs/instances/adding-removing-ssh-keys